+++ canaille/backends/sql/models.py
@@ -179,7 +179,7 @@ class User(canaille.core.models.User, Base, SqlAlchemyModel):
         self.permissions = set()
         self.read = set()
         self.write = set()
-        for access_group_name, details in current_app.config.get("ACL", {}).items():
+        for access_group_name, details in current_app.config["CANAILLE"]["ACL"].items():
             if self.match_filter(details.get("FILTER")):
                 self.permissions |= set(details.get("PERMISSIONS", []))
                 self.read |= set(details.get("READ", []))