+++ canaille/core/endpoints/forms.py
@@ -47,9 +47,9 @@ def unique_group(form, field):
 
 
 def existing_login(form, field):
-    if not current_app.config.get(
-        "HIDE_INVALID_LOGINS", True
-    ) and not models.User.get_from_login(field.data):
+    if not current_app.config["CANAILLE"][
+        "HIDE_INVALID_LOGINS"
+    ] and not models.User.get_from_login(field.data):
         raise wtforms.ValidationError(
             _("The login '{login}' does not exist").format(login=field.data)
         )
@@ -365,7 +365,7 @@ class JoinForm(Form):
     )
 
     def validate_email(form, field):
-        if not current_app.config.get("HIDE_INVALID_LOGINS", True):
+        if not current_app.config["CANAILLE"]["HIDE_INVALID_LOGINS"]:
             unique_email(form, field)