+++ canaille/oidc/oauth.py
@@ -35,23 +35,7 @@ from werkzeug.security import gen_salt
 
 from canaille.app import models
 
-DEFAULT_JWT_KTY = "RSA"
-DEFAULT_JWT_ALG = "RS256"
-DEFAULT_JWT_EXP = 3600
 AUTHORIZATION_CODE_LIFETIME = 84400
-DEFAULT_JWT_MAPPING = {
-    "SUB": "{{ user.user_name }}",
-    "NAME": "{% if user.formatted_name %}{{ user.formatted_name }}{% endif %}",
-    "PHONE_NUMBER": "{% if user.phone_numbers %}{{ user.phone_numbers[0] }}{% endif %}",
-    "EMAIL": "{% if user.preferred_email %}{{ user.preferred_email }}{% endif %}",
-    "GIVEN_NAME": "{% if user.given_name %}{{ user.given_name }}{% endif %}",
-    "FAMILY_NAME": "{% if user.family_name %}{{ user.family_name }}{% endif %}",
-    "PREFERRED_USERNAME": "{% if user.display_name %}{{ user.display_name }}{% endif %}",
-    "LOCALE": "{% if user.preferred_language %}{{ user.preferred_language }}{% endif %}",
-    "ADDRESS": "{% if user.formatted_address %}{{ user.formatted_address }}{% endif %}",
-    "PICTURE": "{% if user.photo %}{{ url_for('core.account.photo', user=user, field='photo', _external=True) }}{% endif %}",
-    "WEBSITE": "{% if user.profile_url %}{{ user.profile_url }}{% endif %}",
-}
 
 
 def oauth_authorization_server():
@@ -121,7 +105,7 @@ def openid_configuration():
         "subject_types_supported": ["pairwise", "public"],
         "id_token_signing_alg_values_supported": ["RS256", "ES256", "HS256"],
         "prompt_values_supported": ["none"]
-        + (["create"] if current_app.config.get("ENABLE_REGISTRATION") else []),
+        + (["create"] if current_app.config["CANAILLE"]["ENABLE_REGISTRATION"] else []),
     }
 
 
@@ -132,8 +116,8 @@ def exists_nonce(nonce, req):
 
 
 def get_issuer():
-    if current_app.config["OIDC"]["JWT"].get("ISS"):
-        return current_app.config["OIDC"]["JWT"].get("ISS")
+    if current_app.config["CANAILLE_OIDC"]["JWT"]["ISS"]:
+        return current_app.config["CANAILLE_OIDC"]["JWT"]["ISS"]
 
     if current_app.config.get("SERVER_NAME"):
         return current_app.config.get("SERVER_NAME")
@@ -143,18 +127,18 @@ def get_issuer():
 
 def get_jwt_config(grant=None):
     return {
-        "key": current_app.config["OIDC"]["JWT"]["PRIVATE_KEY"],
-        "alg": current_app.config["OIDC"]["JWT"].get("ALG", DEFAULT_JWT_ALG),
+        "key": current_app.config["CANAILLE_OIDC"]["JWT"]["PRIVATE_KEY"],
+        "alg": current_app.config["CANAILLE_OIDC"]["JWT"]["ALG"],
         "iss": get_issuer(),
-        "exp": current_app.config["OIDC"]["JWT"].get("EXP", DEFAULT_JWT_EXP),
+        "exp": current_app.config["CANAILLE_OIDC"]["JWT"]["EXP"],
     }
 
 
 def get_jwks():
-    kty = current_app.config["OIDC"]["JWT"].get("KTY", DEFAULT_JWT_KTY)
-    alg = current_app.config["OIDC"]["JWT"].get("ALG", DEFAULT_JWT_ALG)
+    kty = current_app.config["CANAILLE_OIDC"]["JWT"]["KTY"]
+    alg = current_app.config["CANAILLE_OIDC"]["JWT"]["ALG"]
     jwk = JsonWebKey.import_key(
-        current_app.config["OIDC"]["JWT"]["PUBLIC_KEY"], {"kty": kty}
+        current_app.config["CANAILLE_OIDC"]["JWT"]["PUBLIC_KEY"], {"kty": kty}
     )
     return {
         "keys": [
@@ -204,8 +188,7 @@ def generate_user_info(user, scope):
 
 def generate_user_claims(user, claims, jwt_mapping_config=None):
     jwt_mapping_config = {
-        **DEFAULT_JWT_MAPPING,
-        **(current_app.config["OIDC"]["JWT"].get("MAPPING") or {}),
+        **(current_app.config["CANAILLE_OIDC"]["JWT"]["MAPPING"]),
         **(jwt_mapping_config or {}),
     }
 
@@ -420,9 +403,7 @@ class IntrospectionEndpoint(_IntrospectionEndpoint):
 
 class ClientManagementMixin:
     def authenticate_token(self, request):
-        if current_app.config.get("OIDC", {}).get(
-            "DYNAMIC_CLIENT_REGISTRATION_OPEN", False
-        ):
+        if current_app.config["CANAILLE_OIDC"]["DYNAMIC_CLIENT_REGISTRATION_OPEN"]:
             return True
 
         auth_header = request.headers.get("Authorization")
@@ -431,7 +412,7 @@ class ClientManagementMixin:
 
         bearer_token = auth_header.split()[1]
         if bearer_token not in (
-            current_app.config.get("OIDC", {}).get("DYNAMIC_CLIENT_REGISTRATION_TOKENS")
+            current_app.config["CANAILLE_OIDC"]["DYNAMIC_CLIENT_REGISTRATION_TOKENS"]
             or []
         ):
             return None
@@ -445,7 +426,7 @@ class ClientManagementMixin:
     def resolve_public_key(self, request):
         # At the moment the only keypair accepted in software statement
         # is the one used to isues JWTs. This might change somedays.
-        return current_app.config["OIDC"]["JWT"]["PUBLIC_KEY"]
+        return current_app.config["CANAILLE_OIDC"]["JWT"]["PUBLIC_KEY"]
 
     def client_convert_data(self, **kwargs):
         if "client_id_issued_at" in kwargs:
@@ -533,6 +514,9 @@ def generate_access_token(client, grant_type, user, scope):
 
 
 def setup_oauth(app):
+    # hacky, but needed for tests as somehow the same 'authorization' object is used
+    # between tests
+    authorization.__init__()
     authorization.init_app(app, query_client=query_client, save_token=save_token)
 
     authorization.register_grant(PasswordGrant)
@@ -543,9 +527,7 @@ def setup_oauth(app):
     authorization.register_grant(
         AuthorizationCodeGrant,
         [
-            OpenIDCode(
-                require_nonce=app.config.get("OIDC", {}).get("REQUIRE_NONCE", True)
-            ),
+            OpenIDCode(require_nonce=app.config["CANAILLE_OIDC"]["REQUIRE_NONCE"]),
             CodeChallenge(required=True),
         ],
     )