+++ tests/backends/test_backends.py
@@ -10,7 +10,7 @@ def test_required_methods(testclient):
     with pytest.raises(NotImplementedError):
         BaseBackend.validate({})
 
-    backend = BaseBackend(testclient.app.config)
+    backend = BaseBackend(testclient.app.config["CANAILLE"])
 
     with pytest.raises(NotImplementedError):
         backend.has_account_lockability()