+++ tests/oidc/test_configuration.py
@@ -3,18 +3,21 @@ import warnings
 import pytest
 
 from canaille.app.configuration import ConfigurationException
+from canaille.app.configuration import settings_factory
 from canaille.app.configuration import validate
 from canaille.oidc.oauth import get_issuer
 
 
 def test_issuer(testclient):
     with warnings.catch_warnings(record=True):
-        testclient.app.config["OIDC"]["JWT"]["ISS"] = "https://anyauth.mydomain.tld"
+        testclient.app.config["CANAILLE_OIDC"]["JWT"]["ISS"] = (
+            "https://anyauth.mydomain.tld"
+        )
         testclient.app.config["SERVER_NAME"] = "https://otherauth.mydomain.tld"
         with testclient.app.test_request_context("/"):
             assert get_issuer() == "https://anyauth.mydomain.tld"
 
-        del testclient.app.config["OIDC"]["JWT"]["ISS"]
+        testclient.app.config["CANAILLE_OIDC"]["JWT"]["ISS"] = None
         with testclient.app.test_request_context("/"):
             assert get_issuer() == "https://otherauth.mydomain.tld"
 
@@ -24,18 +27,24 @@ def test_issuer(testclient):
 
 
 def test_no_private_key(testclient, configuration):
-    del configuration["OIDC"]["JWT"]["PRIVATE_KEY"]
+    del configuration["CANAILLE_OIDC"]["JWT"]["PRIVATE_KEY"]
+    config_obj = settings_factory(configuration)
+    config_dict = config_obj.model_dump()
+
     with pytest.raises(
         ConfigurationException,
         match=r"No private key has been set",
     ):
-        validate(configuration)
+        validate(config_dict)
 
 
 def test_no_public_key(testclient, configuration):
-    del configuration["OIDC"]["JWT"]["PUBLIC_KEY"]
+    del configuration["CANAILLE_OIDC"]["JWT"]["PUBLIC_KEY"]
+    config_obj = settings_factory(configuration)
+    config_dict = config_obj.model_dump()
+
     with pytest.raises(
         ConfigurationException,
         match=r"No public key has been set",
     ):
-        validate(configuration)
+        validate(config_dict)